requirement.txt
PyYAML
pylint
pytest
pytest-cov
selenium
python-twitter
